now works with pool
