+++ app/core/record/index.ts
@@ -1,0 +1,3 @@
+export * from "./record";
+export * from "./decorators";
+export * from "./errors";
